KAN-72 refactor: 소셜 ID 처리 로직을 Long에서 BigInteger로 변경

+++ src/main/java/com/spring/be/jwt/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.spring.be.jwt.repository;
 import com.spring.be.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
+
 public interface UserRepository extends JpaRepository<User, Long> {
-    User findBySocialId(Long socialId);
+    User findBySocialId(BigInteger socialId);
 }
